restructured repo; broke out cli into it's own file; moved functions that are not cli and if __name__ = '__main__' functions to utils
