+++ json2rst.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python3
+
+import json2rst
