+++ json2rst/pivot.py
@@ -28,8 +28,10 @@ will create an rst file containing the following table:
 +----+-------------+-----------------+
 
 """
+import sys, os
 from typing import List, Dict
 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '.')))
 
 class Pivot:
     def __init__(self,
@@ -43,8 +45,13 @@ class Pivot:
             json_objects (List[Dict[str,str]]): Takes a list
                 of JSON objects and pivots.
         """
+        assert(isinstance(json_objects, list)), \
+            "Pivot only takes a list of JSON objects: {}"
+        for obj in json_objects:
+            assert(isinstance(obj, dict)), \
+                "{} must be a dictionary.".format(obj)
 
-        self.objects = json_objects
+        self.json_objects = json_objects
         self.header_list = header_list
 
     def pivot(self) -> str:
@@ -55,13 +62,16 @@ class Pivot:
 
         return ""
 
-    def _assert_identical_keys(self) -> bool:
+    def _check_pivot_keys(self) -> bool:
         """
         Checks if all important keys are
         available in JSON objects to pivot.
 
         """
-        if not True:
-            return False
+
+        for obj in self.json_objects:
+            for item in self.header_list:
+                assert(item in list(obj.keys())), \
+                    "All dictionaries used in Pivot must have {} key".format(item)
 
         return True
